auto cancel orders

+++ BCPrice_Catcher/Trader/BtccTrader.cs
@@ -199,7 +199,7 @@ namespace BCPrice_Catcher.Trader
 			return null;
 		}
 
-		public override bool CancelPlacedOrder(int orderId, CoinType coinType)
+		public override bool CancelPlacedOrder(long orderId, CoinType coinType)
 		{
 			var builder = new BtccParasTextBuilder("cancelOrder");
 			//price must be added earlier
